Use state value in DropdownButtonFieldForm (#37145)
The current implementation of DropdownButtonFormField does not pass the initial value to _DropdownButtonFormFieldState. As a result changes made through the child DropdownButton are not made to the FormFieldState and the widget is not updated unless a onChanged function is provided to the DropdownButtomFormField constructor.

This change modifies DropdownButtonFormField to behave more consistently with other FormField widgets in how the Form state is handled.